docs(nix/readTree): define "merge" and clarify priority
The readTree documentation refers to "merging" attrsets but doesn't define
this term.  To avoid confusion with the NixOS module merge, let's explain what
this term means.

This commit also makes it more clear that when `default.nix` has a an
attribute `foo` and a sibling directory `foo`, the latter overrules the
former.

Change-Id: Ibe1b8c541a0bf953c963b307e9dc6203349c2a8e