feat(tvix/castore/directory/traverse): use castore Paths
This switches from using std::path::Path to using castore paths.

We can drop some error handlingin descend_to, as absolute (or redundant)
paths are not representable.

We however now need to convert from a std::path::Path to our
representation, and decide how/if we do canonicalization (which depends
on context).

For now this is inline, to be moved to a separate function later.

Change-Id: I5e94693808420c5d56587c68731252b54755bf93