Update patch set 3
Abandoned

With all of the cleanups that happened around fetchers etc, there's luckily not really a reason to do this anymore, especially not bake it into the protocol.

The only place where we still need to do this is when ingesting a single file in recursive mode, inside builtins.file, and we know that case upfront, so can arguably calculate the nar hash in-memory there, if we're concerned about the roundtrip.

Patch-set: 3
Status: abandoned
Tag: autogenerated:gerrit:abandon