feat(tvix/tracing): introduce common tvix-tracing crate
Introduce a new common crate that contains tracing boilerplate which then
can be used in the cli, tvix-store and tvix-build crates.
It has a otlp as optional feature and it has a pub use for the tracing
and indicatif crates so these don't need to be included as dependencies
in the executable crates.

Change-Id: I41468ac4d9c65174515d721513b96fea463d6ed2