fix(tvix/tracing): apply envfilter to all layers
Currently we apply the EnvFilter only to the stderr output writer and
did not affect any other layer, like the otlp layer.
This had the result that we can't control if spans from h2::,
tokio_util:: or other third party dependencies should be exported to
otlp. This currently leads to a lot of h2:: spans with level debug that
are being exported to otlp, with this change our default behavior
changes and debug spans will no longer be exported to otlp, leading to a
lot less spans being exported.

Change-Id: I9f3a7233e9d0aeaa81fe08914579f0b3c80d134e