feat(tvix/tvixbolt): build with crate2nix
This builds tvixbolt with crate2nix, relying on the nixpkgs/crate2nix cross fixes in
  - https://github.com/NixOS/nixpkgs/pull/319153
  - https://github.com/NixOS/nixpkgs/pull/317755
  - https://github.com/nix-community/crate2nix/pull/352

It's meant to be a WIP CL until the changes get merged and land into our pinned nixpkgs/crate2nix sources.

Because crate2nix is not able to use crates built elsewhere in depot, and pick up their src filtering and customized defaultCrateOverrides, tvixbolt was moved into the //tvix workspace for now.

Change-Id: Ie43b08f8c7e67ea527f8213108e6a62c14f70128