docs: move Registration to the top; minor reword
When trying to follow Review documentation, I got tripped up on the
“Gerrit setup” section that was asking about SSH keys since I hadn’t yet
gotten an account setup but the registration setion was near the bottom
of the document despite being a requirement for setting up Gerrit.

Included words specifying SSO, specifies that the TVL account is an
alternative method, & commit→patch since you send a patch with the
commit.

Change-Id: I465c8a79979b368ab01bcebdc7e5ef16e65dd341