feat(tvix/tvix-store): improve progress bar situation
Don't show an empty spinner for daemon commands.
Move the bar to the right, so the text is better aligned between spinner
progress and bar progress styles.

Generally, push progress bars a bit more down to the place where we can
track progress. This includes adding one in the upload_blob span.

Introduce another progress style template for transfers, which
interprets the counter as bytes, not just a plain integer, and use it
for that.

Change-Id: I15c2ea3d2b24b5186cec19cd3dbd706638497f40