feat(tvix/tracing): gRPC trace context propagation
This introduces optional helper function w in tvix/tracing for trace
propagation and uses these helper in the `tvix-store`.
The GRPCBlobService, GRPCDirectoryService and GRPCPathInfoService now
only accept a client with an interceptor because support both would make
the types of these services to complex and the interceptor function can
be a noop if trace propagation is not needed. This could be resolved in
the future if we have a need support for a client without an
interceptor.

This was tested and validated by starting a `tvix-store daemon` and
`tvix-store import`.

Change-Id: I4b194483bf09266820104b4b56e4a135dca2b77a