chore(3p/overlays): remove custom nix_2_3 overlays
Our CI seems to be struggling with using Nix 2.3 a lot, there's various
hangs and unavailable build slots for hours, and based on other testing,
most of these issues seem to be fixed in recent versions of Nix.

This moves `pkgs.nix` to how/what it is in nixpkgs, so whitby is using
whatever pkgs.nix in nixpkgs is using.

People explicitly wanting to keep nix 2.3 can configure it by setting
`nix.package = pkgs.nix_2_3;`.

Change-Id: I2a73337b5ce473862938fc2426c1b74a88b568ea